+++ tests/unit/test_models/test_full_battery_models/test_base_battery_model.py
@@ -388,6 +388,20 @@ class TestBaseBatteryModel(unittest.TestCase):
         model = pybamm.BaseModel()
         self.assertEqual(model.timescale.evaluate(), 1)
 
+    def test_option_type(self):
+        # no entry gets default options
+        model = pybamm.BaseBatteryModel()
+        self.assertIsInstance(model.options, pybamm.BatteryModelOptions)
+
+        # dict options get converted to BatteryModelOptions
+        model = pybamm.BaseBatteryModel({"thermal": "isothermal"})
+        self.assertIsInstance(model.options, pybamm.BatteryModelOptions)
+
+        # special dict types are not changed
+        options = pybamm.FuzzyDict({"thermal": "isothermal"})
+        model = pybamm.BaseBatteryModel(options)
+        self.assertEqual(model.options, options)
+
 
 class TestOptions(unittest.TestCase):
     def test_print_options(self):